pkg/ipmi: * Move constants from ipmi.go to constants.go * Move structures from ipmi.go to structures.go
